requirements.txt
flask
python-dotenv
twilio
google-generativeai